requirements.txt
playwright>=1.58.0
beautifulsoup4>=4.12.3
dateparser>=1.2.0
google-api-python-client>=2.131.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
